requirements.txt
aiogram==2.19
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
Babel==2.9.1
certifi==2021.10.8
charset-normalizer==2.0.11
frozenlist==1.3.0
idna==3.3
multidict==6.0.2
Pillow==9.0.1
pytz==2021.3
requests==2.27.1
urllib3==1.26.8
yarl==1.7.2
